+++ packages/web/components/Landing/Signup/lists.ts
@@ -1,0 +1,128 @@
+export const playerVisitorList = [
+  { perk: 'Access to educational resources', checked: true },
+  { perk: 'Access to community calls', checked: true },
+  { perk: 'List of MetaAlliance guilds', checked: true },
+  { perk: 'Free newsletter', checked: true },
+  { perk: 'Search & filter people', checked: false },
+  { perk: 'Access to the community', checked: false },
+  { perk: 'Ability to earn reputation', checked: false },
+  { perk: 'Ability to earn Seed tokens', checked: false },
+  { perk: 'Get ranked & unlock perks', checked: false },
+];
+
+export const playerMemberList = [
+  { perk: 'Access to educational resources', checked: true },
+  { perk: 'Access to community calls', checked: true },
+  { perk: 'List of MetaAlliance guilds', checked: true },
+  { perk: 'Free newsletter', checked: true },
+  { perk: 'Search & filter people', checked: true },
+  { perk: 'Access to the community', checked: true },
+  { perk: 'Ability to earn reputation', checked: true },
+  { perk: 'Ability to earn Seed tokens', checked: true },
+  { perk: 'Get ranked & unlock perks', checked: true },
+];
+
+export const guildFreeList = [
+  { perk: 'Guild page in MG', checked: true },
+  { perk: 'Members directory', checked: true },
+  { perk: 'Your leaderboard', checked: true },
+  { perk: 'Onboarding paths', checked: true },
+  { perk: 'A guild2guild meetup', checked: true },
+  { perk: 'Your news inside MG', checked: true },
+  { perk: 'Post calls to action', checked: false },
+  { perk: 'Search & filter people', checked: false },
+  { perk: 'We shill your grant', checked: false },
+  { perk: 'X thread about you', checked: false },
+  { perk: 'In the follow list on X', checked: false },
+  { perk: 'Your news in the newsletter', checked: false },
+  { perk: 'A podcast interview', checked: false },
+  { perk: 'Part of MetaAlliance', checked: false },
+  { perk: 'A post about you', checked: false },
+  { perk: 'Branch in The Onboarding Game', checked: false },
+];
+
+export const guildBasicList = [
+  { perk: 'Guild page in MG', checked: true },
+  { perk: 'Members directory', checked: true },
+  { perk: 'Your leaderboard', checked: true },
+  { perk: 'Onboarding paths', checked: true },
+  { perk: 'A guild2guild meetup', checked: true },
+  { perk: 'Your news inside MG', checked: true },
+  { perk: 'Post calls to action', checked: true },
+  { perk: 'Search & filter people', checked: true },
+  { perk: 'We shill your grant', checked: true },
+  { perk: 'X thread about you', checked: true },
+  { perk: 'In the follow list on X', checked: true },
+  { perk: 'Your news in the newsletter', checked: true },
+  { perk: 'A podcast interview', checked: true },
+  { perk: 'Part of MetaAlliance', checked: false },
+  { perk: 'A post about you', checked: false },
+  { perk: 'Branch in The Onboarding Game', checked: false },
+];
+
+export const guildProList = [
+  { perk: 'Guild page in MG', checked: true },
+  { perk: 'Members directory', checked: true },
+  { perk: 'Your leaderboard', checked: true },
+  { perk: 'Onboarding paths', checked: true },
+  { perk: 'A guild2guild meetup', checked: true },
+  { perk: 'Your news inside MG', checked: true },
+  { perk: 'Post calls to action', checked: true },
+  { perk: 'Search & filter people', checked: true },
+  { perk: 'We shill your grant', checked: true },
+  { perk: 'X thread about you', checked: true },
+  { perk: 'In the follow list on X', checked: true },
+  { perk: 'Your news in the newsletter', checked: true },
+  { perk: 'A podcast interview', checked: true },
+  { perk: 'Part of MetaAlliance', checked: true },
+  { perk: 'A post about you', checked: true },
+  { perk: 'Branch in The Onboarding Game', checked: true },
+];
+
+export const patronFreeList = [
+  { perk: 'Access to educational resources', checked: true },
+  { perk: 'Access to community calls', checked: true },
+  { perk: 'List of MetaAlliance guilds', checked: true },
+  { perk: 'Free newsletter', checked: true },
+  { perk: 'Search & filter people', checked: false },
+  { perk: 'Access to the community', checked: false },
+  { perk: 'Ability to earn reputation', checked: false },
+  { perk: 'Ability to earn Seed tokens', checked: false },
+  { perk: 'Get ranked & unlock perks', checked: false },
+];
+
+export const patronMemberList = [
+  { perk: 'Access to educational resources', checked: true },
+  { perk: 'Access to community calls', checked: true },
+  { perk: 'List of MetaAlliance guilds', checked: true },
+  { perk: 'Free newsletter', checked: true },
+  { perk: 'Search & filter people', checked: true },
+  { perk: 'Access to the community', checked: true },
+  { perk: 'Ability to earn reputation', checked: true },
+  { perk: 'Ability to earn Seed tokens', checked: true },
+  { perk: 'Get ranked & unlock perks', checked: true },
+];
+
+export const playerReasons = [
+  'Learn about Web3, DAOs, ReFi & the Metacrisis',
+  'Join the ecosystem & start making friends',
+  'Start accruing reputation (XP) & tokens (Seeds)',
+  'Gain experience & increase chances of landing gigs',
+  'Get help for starting your own project',
+];
+
+export const guildReasons = [
+  'Access a network of builders, early adopters & projects',
+  'Get help or find things you need for your guild',
+  'Profiles & leaderboards for your members',
+  'Gain visibility for your tools or services',
+  'Integrate your tools or services into MetaOS',
+];
+
+export const patronReasons = [
+  'Because you love the idea of MetaGame',
+  'You want to help fix the onboarding problem for the DAO space',
+  'You want to see building DAOs become 10x easier & better',
+  "Besides Web3 & DAOs, you're into ReFi, Game B & Network States",
+  'Membership & platform utility will be paid in Seeds',
+];
